+++ CHANGELOG.md
@@ -2,10 +2,19 @@
 
 ## Development
 
+### Bugs fixed
+* Partially fixed handling of compound intervals in reference annotations (#46)
+* Fixed issue with retrieving upstream context for minus strand genes near contig ends
+* Resolved issues involving reference annotations with multiple contigs/chromosomes (#48)
+
 ### Enhancements
 * Output GFF files no longer include the genome sequence.
 
 ### Housekeeping
+* merged genes notes:
+  - removed redundant note text indicating the annotation source
+  - include both locus tag and gene name of affected genes
+  - always make sure the word "merge" is included in the note for consistency
 * Removed largely duplicated (and slightly buggy) code for comparing RATT
   and ab initio annotations